refactor: split input processing
